ClaudeTest/requirements.txt
# Web Framework
streamlit==1.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pandasql==0.7.3

# AI/LLM
anthropic==0.18.1

# Visualization
plotly==5.18.0

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Authentication
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2
